backend/SmartLockSystemServer/requirements.txt
httpx>=0.26.0
fastapi>=0.109.0
starlette>=0.35.1
uvicorn>=0.26.0
opencv-python>=4.9.0.80
tensorflow>=2.15.0
numpy>=1.26.3
paho-mqtt>=1.6.1
matplotlib>=3.8.2
keras~=2.15.0
scipy~=1.12.0
pillow~=10.2.0
scikit-learn~=1.4.0
sqlalchemy
PyJWT
psycopg2
pydantic-settings
passlib